requirements.txt
click==4.1
python-daemon==2.0.5
redis==2.10.3
